BST-4 start Rust implementation
Rust is gaining in popularity as a way to do better
systems programming, where memory management is explicit
at compile time.

This commit is the start of a Rust implementation. The goal
is getting the pipeline hooked up, and an initial implementation
of minimal functionality: instantiate a node, and ensure it is
testable.