BST-23 stern-brocot - initial implementation (#25)
integers >= 0 work

This turned out to be a bit easier than I
thought it would be, for at least the
simple (trivial) iterative technique. Hopefully
I got it right and it will work for rational
numbers, because that's the point.

rational numbers <= 1

This is partially working, committed to keep
track of where the implementation needs to be
fixed. Whatever the problem is, it's almost
surely very simple.