BST-24 initial implementation of AVL tree (#27)
This massive commit message is the result of squashing
over a dozen commits. The original intent was to merge
the branch after correctly implementing AVL tree. Sadly,
a correct implementation is not quite there. To keep
everything efficient and reduce the need for rebasing,
the AVL branch is merged in its incorrect state. Hopefully,
this will make it easier to work on going forward.

avl: rotation spec clarifications

Mostly added a lot of context blocks around
specific rotation specs to better describe
which tree (node arrangement) is being tested.
The existing rotation code seems to work correctly.

avl: start of balance specs

Balancing by retracing is the next step. This commit
provides an initial stab at what the tests should look
like. I have other things to do today, will fill remaining
tests and implementation later.

There is an overhead associated with computing the
height on demand. I hesitate to memoize it as the
tree will change height at any node over its
lifespan.

avl: more example specs

Along with adding more examples, there is a marker
for working through the mechanics of the recursion
on insert. It's pretty interesting, and not exactly
how I though it would work. Could be as much as a
day's work to go through how the stack works with
respect to arguments passed into the recursing
method, which is overloaded, btw, we're calling super,
that's the interesting part.

avl: placeholder commit, need to refactor spec on master

The insertion for avl needs to move from node to tree,
which really needs to leverage some refactoring of
the specs, which should really be done in master.
So committing here and hopefully the rebase won't
collide with any of the code already on this branch.

avl: initial rebalancing for left chain

Seems to work correctly. I tried to implement
using the code on Wikipedia as a template, but
that code is far too messy for me to understand.
So I'm working it out on my own on paper for
special cases, and implementing by these cases.

avl: naming clean up for semantic congruence

Changed several method names in the rebalancing
procedure to better reflect what the methods are doing.

Specifically, "rebalance" is now "retrace," and only
traverses up the parent pointers, invoking the "balance"
method each iteration. Much easier to understand. At
least for me.

avl: so-called double rotation implemented

The double rotation operation is very simple to
implement by composing the two single rotations
opposing each other. The key is using the correct
rotation roots for each.

avl: cleaned up naming convention for rotation

The names for the rotation methods are now very
simple: rotate_[left|right[_[right|left]]. An
alternative scheme would have been to stick
with clockwise and counterclockwise, but the
abbreviations "cw" and "ccw" aren't descriptive
enough, and spelling it out makes for very long
method names.

avl: "knee" rotation case from tree retracing

I think this is the commit which finishes the tree
insertion which handles retracing and rebalancing
correctly.

The main thing I've gotten out of this is that the
references I've been using make it more difficult
than it really is, primarily because they all use
large functions which do too much.

There is also the notion of returning an updated
root node from a rotation. This may be a useful notion,
and worth investigating in more detail.