BST-44 create a yaml fixture directory at repo root (#38)
From the README:

Normally, test fixtures would go into a directory
structure local to a specific implementation. In this
case, a set of fixtures common to all implementations
is useful.

This commit uses the Ruby implementation as an initial
proof of concept. The Ruby tree Generator class was used
to create the initial yaml fixtures, and all of the example
trees defined in the Generator class are written to yaml.
Some modifications had to be made to allow a pre-specified
uuid to be defined for the nodes, otherwise it's really
hard to test nodes by comparison when the uuid is truly
random.

Note that Ruby has a built-in `to_yaml` method for objects.
The decision here is to not override that method, but
instead implement `to_yml` to store just the data for each
tree, without any of the Ruby metadata associated with
the Ruby built-in `to_yaml`.