BST-77 javascript invert binary tree
Inverting according to definition given on Leetcode,
which is to swap left and right subtrees at every node.

Scoping `this` becaome an issue, hence the update to
the post order traverse call back. Commented out console
logging is left in for now, will be removed in a future
commit.

The lack of standard examples and tooling for property-
based testing slowed this work down. All languages
implementations need this.