BST-150 fix double AVL rotations (#82)
The node spec is fully working as of this commit. It turns
out my logic was mostly fine. Fixing a parent pointer
assignment in the previous commit along with fixing a
copy pasta error in this commit got everything working
the way it's supposed to. I recall the last time I worked
on this I felt like I was going blind trying to figure
out what was wrong.