BST-151 fix the AVL tree retrace method
For the actual #retrace method, it was adjusting
the rotate_left to match rotate_right, as called
from retrace. Not much modification.

Most of the changes are in the avl_tree_spec, for
which the retrace spec for left children was restructured
extensively to enable easier comprehension of the
spec and the algorithm.

It would be good to set up some regression testing
for both rotate_left and rotate_right, but that will
have to wait for a future commit. Had these two methods
been thoroughly tested, the retrace specs would probably
be quite a bit shorter.