BST-152 fix avl perfect tree spec (#84)
A full fix did not happen in this commit. However,
the spec in question was cleaned up a bit, as was
several other parts of the code, namely removing
stale comments.

Getting the spec to work will likely require rewriting
of the rotation and balancing implementation, which could
be really tedious. Hopefully, the pattern established in
this change will make those reimplementations a bit easier
and more reliable.