BST-173 semaphore configuration for python (#97)
Initial configuration for Python works, but the
dependency configuration could almost surely be improved.
For some reason, the pytest block doesn't seem to be
using the configuration provided by python set up
block, and it's not clear what the problem is. Digging
into the efficiency here is pointless right now as it's
building fast enough. In a production system, with a much
larger build, it would probably be worth digging in a
lot deeper to understand what semaphore is and is not
doing for python caching.