BST-180 Create fixtures for full BSTs
In the quest to implement algorithm for determining all
properties and operations, this change provides a common
set of fixtures for full binary search trees described in
yaml, using the defined from the cited references.

Some really interesting structure emerges when the consective
integers are sorted into a binary search tree in a particular.
More details in a later commit.