BST-47 libyaml testing structure (#153)
The event stream API for libyaml was chosen to implement
a binary search tree configuration loader. Since c doesn't
have any native libraries for working with associative arrays,
a purpose-built struct is defined and passed into the event
loop to keep track of state.

As event stream is linear, the event processor must keep
track of the relative position in the tree. This is accomplished
with a pointer to the currently processing node (current_node),
on an implicit stack. Push and pop are handled via pointer
swaps with the appropriate child and its parent. The root_node
pointer is a convenience, and is not strictly necessary.

A rudimentary test suite is implemented which confirms the
size, height, root node key, and validity of the processed
tree.