BST-237 Compute balance factor for an AvlNode (#162)
Most of the existing AVL implementation is bottom-up, focused
on the rotations, all of which work correctly. But the overall
AVL algorithm does not work correctly.

I've tried to work around explicitly computing a balance factor as it
requires two calls to height. However, the AvlTree has stymied me for
too long, and the published work on it shows balance factors being
explicitly computed. So I should do it here and rebase the AVL on this
once it's merged.