BST-242 add yard documentation processing
Adding a documentation processor is overdue. Ruby `yard`
processor was chosen as it allows RDoc for backward
compatibility, but otherwise resembles most other documentation
processors syntactically. In the past I've been add a lot of
the documentation and implementation notes to a LaTeX file,
which is inefficient from the standpoint of looking things up.
Independent documentation is convenient for spanning multiple
source code implementations, at least for high level concepts.