BST-133 github action for c
This is a first cut at the problem. Note that my autotools expertise
is likely 10-15 years out of date. However, what is in this file works
for me locally, and works on the github workflow CI. If there is a better
way to do it (and there _has_ to be), let me know.

Specifically, `autoreconf` may be able to eliminate a lot of the manual
steps, and will be the subject of a future investigation.