BST-256 print a single node to console
This change is part of supporting yaml definitions for binary search
trees which are in the `fixtures` directory. To correctly load a
bst fixture, the yaml needs to be processed in order, that is,
reading the key and instantiate a node, then traverse left, then
traverse right. The c++ implementation lack pre_order traversal
at the moment, so the implementation is blocked. Printing to console
is a utility for helping define a pre_order traversal, it will be
the first callback function.

I'm back working full time, and don't have a lot of time to
work through the details of c++ in long sessions.