BST-269 remove automatically committing diagram (#190)
The repo visualizer diagram makes an automatic commit and attaches it to
my pull requests, which I find absolutely infuriating. The pull request
which is emitted also fails the title checking action.

I would like to have this capability, but it's too much
trouble for too little payback right now.