BST-311 Stern-Brocot implementation update
Doing this right has to deal with all the vagaries of base and
representation in binary of number in that base. Using base 10 as
default, even numbers such as 0.33 have more nuance than I expected.
Going further on the topic would require some really focused study, for
which I do not have the energy at the time of writing.