BST-327 Ruby diameter implementation
https://doolin.atlassian.net/browse/BST-327

Diameter implementation is relatively straightforward.
When I started the implementation, I had a notion
of what I wanted, and as it turns out, Copilot
suggested something very close to that, which I
went with. Copilot's suggestion was incorrect
with respect to initial values, but the overall
methodology was iteratively better than what I
had in mind. It took a few minutes to adjust
the suggested code, but it's good to go.